bot/requirements.txt
python-telegram-bot==20.0a0
SQLAlchemy==1.4.37
PyMySQL==1.0.2
python-dotenv==0.20.0